Version 1.1.0
Rewrite to better handle unit switching between imperial and metric.
Instead of re-writing the node definition and trying to dynamically
modify the Polyglot internal drivers array, create nodes status values
that use mulit-uom editors. When a weather data value is sent to the
ISY, send it in the user's prefered units.  The ISY will then display
it properly and also handle it properly in programs.