Clean up forecast data.
OWM forecast data is different from other service providers in
that they don't provide daily forecasts but instead provide 40 3-hour
forecast blocks.  These will start with the next 3 hour block that
is after the query time.  Thus the first and last day's may be
incomplete views of the forecast. As an added complexity, the blocks
are based on UTC time so one 3 hour block may straddle two different
days when converted to the local timezone.

The node server tries to hide all this from the user and displays
daily forecast nodes.  The data for each daily forecast is derrived
from the 8 3hr blocks (using local timezone) that start within each
day. If 8 blocks are not available, that node is not updated.  The
result is such that forecast_0 should get updated at the beginning
of the day and will then remain static throughout the day.  The
remaining forecast nodes will update as new data becomes available.

The above changes then exposed another problem. If the node server
was restarted in the middle of the day, the forecast_0 node would
revert to it's default UOMs even though the actual values don't
change. Since we don't update the node, it would stay that way
with the incorrect UOMs until the next day when the node was
updated.  To fix this, the daily node will now dynamically build
its drivers array using the user's preferred UOM so that restarts
don't result in the UOM changing.